pyproject.toml
[project]
name = "tgspin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=0.21.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "cachetools>=5.3.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.25.0",
    "psycopg2-binary>=2.9.9",
]
